Limit search to one repo in StalledDiscussionFlow
Before this change, trying to test this flow against a fork would lead
to the initial search finding results in my fork + the bf2 repo +
Tom's repo. Then when it would try to add a label to a PR anywhere
other than my fork, it would throw an exception.

There doesn't seem to be an easy way to figure out which repo
the app installation is for from the client, so making it a config
property seems the most straightforward.