Update to newer quarkus(-github-app) for features
We'll need this to be able to use the ConfigFileProvider in the
StalledDiscussionFlow, since the @ConfigFile annotation only seems to
work well for injection into event-listener methods.

The `org.jetbrains:annotations` dependency doesn't get pulled in
implicitly any longer, so I've just switched usage of the @NotNull &&
@Nullable annotations from there to the SmallRye equivalents, which
are available.

Something that I'm not 100% sure about is the test changes here: I
guess the Jackson/SnakeYAML handling of null fields changed somewhere
along the way.